+++ packages/datadog_session_replay/lib/src/datadog_session_replay.dart
@@ -58,6 +58,8 @@ class DatadogSessionReplay {
             imagePrivacyLevel: _configuration.imagePrivacyLevel,
           ),
           touchPrivacyLevel: _configuration.touchPrivacyLevel,
+          iconRasterLogicalSize: _configuration.iconRasterLogicalSize,
+          internalLogger: internalLogger,
         );
 
   void addElement(Key key, Element e) {